backend/app/core/config.py
import os
from pydantic import BaseModel
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

class Settings(BaseModel):
    PROJECT_NAME: str = "Xue Email Assistant"
    API_V1_STR: str = "/api/v1"
    SECRET_KEY: str = os.getenv("SECRET_KEY", "your-secret-key-for-jwt")
    ALGORITHM: str = "HS256"
    ACCESS_TOKEN_EXPIRE_MINUTES: int = 60 * 24 * 7  # 7 days
    
    # Database
    DATABASE_URL: str = os.getenv("DATABASE_URL", "sqlite:///./app.db")
    
    # CORS
    CORS_ORIGINS: list = ["http://localhost:5173", "http://localhost:8080"]
    
    # First user is admin
    FIRST_USER_IS_ADMIN: bool = True

settings = Settings() 